pyproject.toml
[tool.poetry]
name = "python-telegram-wikibot"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.6"
wikipedia = "^1.4.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
